// src/api/generated/model/instructorDashboardAttentionItemType.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */

/**
 * Category of the attention item
 */
export type InstructorDashboardAttentionItemType = typeof InstructorDashboardAttentionItemType[keyof typeof InstructorDashboardAttentionItemType];


export const InstructorDashboardAttentionItemType = {
  TEAM_APPROVAL: 'TEAM_APPROVAL',
  MISSING_QUESTIONS: 'MISSING_QUESTIONS',
  PENDING_EVALUATION: 'PENDING_EVALUATION',
  RESULTS_READY: 'RESULTS_READY',
} as const;